refactor(new-edition-workspace): limit export scope to only new-edition-workspace
