+++ packages/geo/src/lib/workspace/index.ts
@@ -5,3 +5,4 @@ export * from './workspace-selector/workspace-selector.module';
 export * from './workspace-updator/workspace-updator.directive';
 export * from './workspace-updator/workspace-updator.module';
 export * from './widgets/index';
+export * from './new-edition-workspace/new-edition-workspace';
