+++ packages/geo/src/lib/workspace/shared/index.ts
@@ -5,5 +5,4 @@ export * from './feature-workspace.service';
 export * from './edition-workspace';
 export * from './edition-workspace.service';
 export * from './workspace.utils';
-export * from './new-edition-workspace/index';
 export * from './layer-workspace';
