+++ packages/geo/src/lib/workspace/shared/layer-workspace.ts
@@ -1,9 +1,5 @@
-import {
-  EditionWorkspace,
-  FeatureWorkspace,
-  NewEditionWorkspace,
-  WfsWorkspace
-} from './';
+import { NewEditionWorkspace } from '../new-edition-workspace/new-edition-workspace';
+import { EditionWorkspace, FeatureWorkspace, WfsWorkspace } from './';
 
 export type LayerWorkspace =
   | FeatureWorkspace
